Refactor Order class
